+++ spotify-ui/src/components/playback/sdk/index.ts
@@ -1,0 +1,7 @@
+import initPlaybackSDK from "./initPlaybackSDK";
+import playerEventHandler from "./playerEventHandler";
+
+export {
+  initPlaybackSDK as default,
+  playerEventHandler
+}
