+++ spotify-ui/src/components/playback/sdk/playerEventHandler.ts
@@ -1,0 +1,5 @@
+const playerEventHandler = () => {
+
+}
+
+export default playerEventHandler;
